requirements.txt
# Core dependencies
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0
tabulate>=0.9.0

# LLM API clients
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0
cohere>=4.0.0

# Code analysis and execution
ast-unparse>=1.6.3
pylint>=2.17.0
black>=23.0.0
bandit>=1.7.5
psutil>=5.9.0

# Data handling and reporting
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jinja2>=3.1.0

# Testing and utilities
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
tenacity>=8.2.0
tqdm>=4.65.0

# Optional: For Hugging Face models
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0